# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: GPS Config & Firmware Center API
  description: >
    โครง API เริ่มต้นสำหรับ Sprint 0 — ยังไม่ครบทุก Endpoint
    เพิ่มเติมทีละ Phase ตามที่ Build จริง (ดู ../planning/01_GPS_Build_Reference.md Section 3)
  version: "0.1.0"
servers:
  - url: http://localhost:3001/api/v1
    description: Local dev

# บังคับ Bearer JWT ทุก endpoint โดย default — override เป็น security: [] เฉพาะ endpoint ที่เป็น public
security:
  - bearerAuth: []
...
paths:
...
  /firmware/{firmwareId}/approve:
    post:
      tags: [firmware]
      operationId: approveFirmware
      summary: QAEngineer อนุมัติคุณภาพ Firmware (Firmware Approval Lifecycle — docs/13 §3.2)
      description: >-
        resource แยกเป็น `firmware-decision` action `Approve` (mirror
        `config-decision`) — QAEngineer เท่านั้น · สถานะ `pending_review` →
        `approved` เท่านั้น (ต้องดูผล `simulateFirmware` มาก่อนตามดุลพินิจ QA
        — ไม่มีขั้น "ส่งต่อ" แยกแบบ Config เพราะ QA เป็นผู้ตัดสินใจคนเดียว)
      parameters:
        - in: path
          name: firmwareId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: อนุมัติสำเร็จ — approvalStatus เป็น approved พร้อม approvedBy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Firmware"
        "401":
          description: Token ไม่ถูกต้องหรือหมดอายุ
        "403":
          description: ไม่มีสิทธิ์อนุมัติคุณภาพ Firmware — เฉพาะ Role QAEngineer เท่านั้น
        "404":
          description: ไม่พบ Firmware นี้
        "409":
          description: approvalStatus ของ Firmware ปัจจุบันไม่ใช่ pending_review จึงอนุมัติไม่ได้
  /firmware/{firmwareId}/reject:
    post:
      tags: [firmware]
      operationId: rejectFirmware
      summary: QAEngineer ปฏิเสธคุณภาพ Firmware (Firmware Approval Lifecycle — docs/13 §3.2)
      description: >-
        resource เดียวกับ `approveFirmware` (`firmware-decision` action
        `Approve`) — สถานะ `pending_review` → `rejected` เท่านั้น · ไม่มี body
        (ยังไม่รับเหตุผล — mirror `rejectConfig`) FirmwareEngineer ต้องอัปโหลด
        เวอร์ชันใหม่แก้ไข ไม่มีการแก้ไฟล์เดิมซ้ำ
      parameters:
        - in: path
          name: firmwareId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: ปฏิเสธสำเร็จ — approvalStatus เป็น rejected (approvedBy ยังเป็น null)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Firmware"
        "401":
          description: Token ไม่ถูกต้องหรือหมดอายุ
        "403":
          description: ไม่มีสิทธิ์ปฏิเสธคุณภาพ Firmware — เฉพาะ Role QAEngineer เท่านั้น
        "404":
          description: ไม่พบ Firmware นี้
        "409":
          description: approvalStatus ของ Firmware ปัจจุบันไม่ใช่ pending_review จึงปฏิเสธไม่ได้
...
  /devices/{deviceId}:
    get:
      tags: [device]
      operationId: getDevice
      summary: ดูรายละเอียดอุปกรณ์ 1 เครื่อง (Device Detail — v3.17)
      description: >-
        อ้างด้วย `Device.deviceId` (เลขเครื่องจริง) ไม่ใช่ `Device.id` (UUID
        ภายใน) · คืนเฉพาะ field ของ record Device เอง — สถานะ Config/Firmware
        ของกล่อง (`GET /devices/{deviceId}/status`) ยังไม่ implement
      parameters:
        - in: path
          name: deviceId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: ข้อมูล Device
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
        "401":
          description: Token ไม่ถูกต้องหรือหมดอายุ
        "404":
          description: ไม่พบ Device (deviceId) นี้
...
  /device-firmware-overrides:
    get:
      tags: [device]
      operationId: listDeviceFirmwareOverrides
      summary: Operation ดูรายการคำขอ Firmware Override ทั้งหมด (Sprint 3 แถวที่ 24, v3.43)
      description: >-
        คิวอนุมัติของ Operation — filter ตาม `status` (optional, ไม่ส่ง = ทุก
        สถานะ) เรียงคำขอใหม่สุดก่อน (`overriddenAt` desc) resource
        `device-firmware-override` action `Read` — Operation เท่านั้น
      parameters:
        - in: query
          name: status
          required: false
          schema:
            type: string
            enum: [pending, approved, rejected]
      responses:
        "200":
          description: รายการคำขอ (เรียงตาม overriddenAt desc)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DeviceFirmwareOverride"
        "401":
          description: Token ไม่ถูกต้องหรือหมดอายุ
        "403":
          description: ไม่มีสิทธิ์ดูคิวอนุมัติ override — เฉพาะ Role Operation เท่านั้น
  /device-firmware-overrides/{id}/approve:
    post:
      tags: [device]
      operationId: approveDeviceFirmwareOverride
      summary: Operation อนุมัติคำขอ Firmware Override (Sprint 3 แถวที่ 24, v3.43)
      description: >-
        เปลี่ยนสถานะคำขอเป็น `approved` เท่านั้น **ไม่ apply เข้าอุปกรณ์ให้
        อัตโนมัติ** ปลดล็อกแค่สิทธิ์ให้ `confirmFirmwareInstall` ยอมรับ
        firmware นี้กับอุปกรณ์เครื่องนี้ได้ resource `device-firmware-override`
        action `Approve` — Operation เท่านั้น (Separation of Duty — ST ผู้ส่ง
        คำขอเองอนุมัติไม่ได้)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: "อนุมัติสำเร็จ — คืนคำขอที่อัปเดตแล้ว (status: approved)"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceFirmwareOverride"
        "401":
          description: Token ไม่ถูกต้องหรือหมดอายุ
        "403":
          description: ไม่มีสิทธิ์อนุมัติ override — เฉพาะ Role Operation เท่านั้น
        "404":
          description: ไม่พบคำขอ override นี้
        "409":
          description: คำขอนี้ถูกตัดสินใจไปแล้ว (approved/rejected) — กันตัดสินใจซ้ำ
  /device-firmware-overrides/{id}/reject:
    post:
      tags: [device]
      operationId: rejectDeviceFirmwareOverride
      summary: Operation ปฏิเสธคำขอ Firmware Override (Sprint 3 แถวที่ 24, v3.43)
      description: >-
        เปลี่ยนสถานะคำขอเป็น `rejected` — `rejectReason` ไม่บังคับ resource
        เดียวกับ approve (action `Approve`)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                rejectReason:
                  type: string
                  maxLength: 500
                  description: เหตุผลที่ปฏิเสธ (optional)
      responses:
        "200":
          description: "ปฏิเสธสำเร็จ — คืนคำขอที่อัปเดตแล้ว (status: rejected)"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceFirmwareOverride"
        "401":
          description: Token ไม่ถูกต้องหรือหมดอายุ
        "403":
          description: ไม่มีสิทธิ์ปฏิเสธ override — เฉพาะ Role Operation เท่านั้น
        "404":
          description: ไม่พบคำขอ override นี้
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    DeviceFirmwareOverride:
      type: object
      description: >-
        Firmware Override รายเครื่อง (Sprint 3 แถวที่ 24) — คำขอให้อุปกรณ์
        เครื่องเดียวติดตั้ง Firmware ตัวที่ระบุได้ แม้ไม่ตรงกับ Campaign
        Rollout ที่กำหนดไว้ ต้องผ่าน Operation อนุมัติก่อนถึงมีผล mirror
        `DeviceConfigOverride` ต่างกันที่ไม่มี `fields`/`configId` เพราะ
        Firmware เป็นเวอร์ชันเดียวทั้งก้อน ไม่ใช่ key-value (ดู comment เหนือ
        `model DeviceFirmwareOverride` ใน schema.prisma)
      required:
        [
          id,
          deviceId,
          firmwareId,
          versionNumber,
          reason,
          status,
          overriddenBy,
          overriddenAt,
        ]
      properties:
        id:
          type: string
          format: uuid
        deviceId:
          type: string
        firmwareId:
          type: string
          format: uuid
        versionNumber:
          type: integer
          description: นับจากทุกสถานะ (รวม rejected) เพิ่มทีละ 1 ต่ออุปกรณ์
        reason:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected]
        overriddenBy:
          type: string
          format: uuid
          description: userId ของ ST ที่ส่งคำขอ
        overriddenAt:
          type: string
          format: date-time
        decidedBy:
          type: string
          format: uuid
          nullable: true
          description: userId ของ Operation ที่อนุมัติ/ปฏิเสธ — null ถ้ายัง pending
        decidedAt:
          type: string
          format: date-time
          nullable: true
        rejectReason:
          type: string
          nullable: true
          description: เหตุผลที่ปฏิเสธ — optional แม้สถานะเป็น rejected
...
    Firmware:
      type: object
      description: >-
        1 ไฟล์ Firmware ที่อัปโหลดเข้าระบบ (v3.23) — `objectKey`/
        `originalFilename`/`fileSizeBytes` ปิด gap ที่ `01_GPS_Build_Reference.md`
        §3 ระบุไว้แต่แรกว่าต้องมี "path ใน Object Storage" แต่ schema เดิม
        (ก่อน Sprint 3 #23) ไม่มี field เก็บ reference ไฟล์เลยสักตัว ·
        `approvalStatus`/`approvedBy` เพิ่มโดย Firmware Approval Lifecycle
        (docs/13_Role_Redesign_Proposal.md §3.2 — PR ถัดจาก #178)
      required:
        [
          id,
          version,
          deviceModelCompatibility,
          uploadStatus,
          deviceUpdateStatus,
          objectKey,
          originalFilename,
          fileSizeBytes,
          uploadedBy,
          uploadedAt,
          approvalStatus,
          approvedBy,
        ]
      properties:
        id:
          type: string
          format: uuid
        version:
          type: string
        deviceModelCompatibility:
          type: array
          items:
            type: string
        uploadStatus:
          type: string
          description: สถานะอัปโหลด/จัดเก็บฝั่งเรา (v3.7)
          enum: [pending, stored, failed]
        deviceUpdateStatus:
          type: string
          description: >-
            สถานะอัปเดตเวอร์ชันจริงของกล่อง อ้างอิงจาก device-status (v3.7) —
            ยังคงเป็น `unknown` เสมอในเวอร์ชันนี้ (v3.23) เพราะ device-status
            ยังไม่ implement (ดู RBAC_Matrix.md Section 6)
          enum: [unknown, up_to_date, pending_update]
        objectKey:
          type: string
          description: >-
            key/path ใน Object Storage bucket (ภายใน — ไม่มี endpoint ให้
            download ไฟล์กลับตอนนี้ เก็บไว้เผื่ออนาคต)
        originalFilename:
          type: string
        fileSizeBytes:
          type: integer
        uploadedBy:
          type: string
          format: uuid
        uploadedAt:
          type: string
          format: date-time
        approvalStatus:
          type: string
          description: >-
            สถานะอนุมัติคุณภาพ (Firmware Approval Lifecycle, docs/13 §3.2) —
            คนละมิติกับ uploadStatus ด้านบน (uploadStatus = ผลทางเทคนิคของ
            การอัปโหลด, approvalStatus = QAEngineer ตรวจคุณภาพแล้วหรือยัง)
            ต้อง `approved` ควบคู่กับ uploadStatus `stored` ถึงใช้สร้าง
            Campaign ได้ (`createCampaign`)
          enum: [pending_review, approved, rejected]
        approvedBy:
          type: string
          format: uuid
          nullable: true
          description: >-
            user id ของ QAEngineer ที่อนุมัติ — null จนกว่าจะถูกอนุมัติ
            (ตอน reject ก็ยังเป็น null เช่นกัน — ไม่มีใคร "อนุมัติ" การ reject)
...
    Device:
      type: object
      description: >-
        อุปกรณ์ GPS หนึ่งเครื่อง (ตรงกับ Prisma model Device — MVP ตาม
        ../planning/03_GPS_Detailed_Build_Steps.md Phase 0 ข้อ 8 · ยังไม่มี FK
        ไป Vehicle/HardwareRevision ตาม Data Dictionary (`customer` เพิ่มแล้ว
        — docs/12_CustomerScope_Proposal.md เฟส B, PR #127/#153) — ดู
        schema.prisma comment)
      required:
        [id, deviceId, simNumber, deviceModel, protocol, status, registeredAt]
      properties:
        id:
          type: string
          format: uuid
          description: surrogate key ภายใน (ใช้ทำ FK/index) — client อ้างอิงด้วย deviceId
        deviceId:
          type: string
          description: เลขเครื่องจริงที่ช่างกรอก/สแกนตอนลงทะเบียน (unique)
          example: DEV-0001
        simNumber:
          type: string
        deviceModel:
          type: string
          example: GT06N
        protocol:
          type: string
          example: TCP
        status:
          type: string
          enum: [registered, installed, decommissioned]
        registeredAt:
          type: string
          format: date-time
        installedAt:
          type: string
          format: date-time
          nullable: true
        customer:
          type: object
          allOf:
            - $ref: "#/components/schemas/CustomerSummary"
          nullable: true
          description: >-
            ลูกค้าที่ผูกกับอุปกรณ์นี้ (ถ้ามี) — null = ยังไม่ได้กำหนดลูกค้า
            รอบนี้เป็น read-only เท่านั้น ยังไม่มี endpoint กำหนด/แก้ลูกค้า
            ของอุปกรณ์ (ดู docs/12 เฟส B ส่วน "ยังไม่ทำ")
    CustomerSummary:
      type: object
      description: >-
        ข้อมูลลูกค้าแบบย่อ (listCustomers, embed ใน Device) — id + ชื่อบริษัท
        เท่านั้น ไม่มี contactName/email/phone/priorityTier (docs/12
        เฟส B, PR #127/#153) — ทุก role ที่ login แล้วเรียกได้ ไม่ควรคืน
        ข้อมูลติดต่อของลูกค้าที่ละเอียดกว่านี้
      required: [id, companyName]
      properties:
        id:
          type: string
          format: uuid
        companyName:
          type: string
          example: ABC Logistics
...